// src/types/global.d.ts
export interface TokenPair {
    blockTimestampLast: bigint,
    creator: Principal
    id: string,
    kLast: bigint,
    lptoken: string,
    price0CumulativeLast: bigint,
    price1CumulativeLast: bigint,
    reserve0: bigint,
    reserve1: bigint
    token0: string,
    token1: string
    totalSupply: bigint,
    lptokens?:string
}

export interface Token {
    index: BigInt;
    canisterId: Principal;
    name: string;
    symbol: string;
    logo: string;
    decimals: BigInt;
    totalSupply: BigInt;
    mintable: boolean;
    burnable: boolean;
    owner: Principal;
    timestamp: BigInt;
}

export interface TokenValue {
    decimals: number;
    fee: bigint;
    id: string
    name: string
    symbol: string
    totalSupply: number,
    logo?: string
}